this was really bad

+++ Quote.py
@@ -1,7 +1,6 @@
 import logging
 import os
 import sqlite3
-
 from errbot import botcmd, BotPlugin
 
 
@@ -14,7 +13,7 @@ class Quote(BotPlugin):
             logging.warning('no database found, creating a new one')
             open(QUOTE_DB, 'a').close()
 
-        self.con = sqlite3.connect('QUOTE_DB', check_same_thread=False)
+        self.con = sqlite3.connect(QUOTE_DB, check_same_thread=False)
         self.cur = self.con.cursor()
         self.cur.execute('''create table if not exists quotes (id integer primary key, quote text not null)''')
         self.con.commit()
@@ -29,51 +28,53 @@ class Quote(BotPlugin):
         """ Returns random quote, usage: !quote"""
         self.cur.execute('''select * from quotes order by random() limit 1''')
         quote = self.cur.fetchone()
-        if quote is None:
-            msg = 'No quotes added yet'
+        if quote is not None:
+            return '[%d] %s' % (quote[0], quote[1])
         else:
-            msg = '[%d] %s' % (quote[0], quote[1])
-        self.con.commit()
-        return msg
+            return 'No quotes added yet'
 
     @botcmd()
     def quote_find(self, msg, args):
         """Searches quotes for strings, usage: !quote find <args>"""
+        if len(args) != 1:
+            return "Usage: !quote find <args>"
+
         self.cur.execute('''select * from quotes where quote like ? order by random() limit 1''', ('%' + args + '%',))
         quote = self.cur.fetchone()
         if quote is None:
-            ret = 'Did not found anything matching = %s.' % (args)
+            return 'Did not found anything matching = %s.' % (args)
         else:
-            ret = '[%d] %s' % (quote[0], quote[1])
-        self.con.commit()
-        return ret
+            return '[%d] %s' % (quote[0], quote[1])
 
     @botcmd()
     def quote_add(self, msg, args):
         """Adds a new quote, usage: !quote add <string> """
+
         self.cur.execute('''insert into quotes (quote) values (?)''', (args,))
         self.con.commit()
 
-        ret = 'Added: %s.' % (args)
-        return ret
+        return 'Added: %s.' % (args)
 
     @botcmd()
     def quote_del(self, msg, args):
         """Removes Quote from Database, usage: !quote del <id>"""
+        if len(args) != 1:
+            return "Usage: !quote del <id>"
+
         self.cur.execute('''delete from quotes where id = ?''', (args,))
         self.con.commit()
 
-        ret = 'Removed: #%d.' % int(args)
-        return ret
+        return 'Removed: #%d.' % int(args)
 
     @botcmd()
     def quote_get(self, msg, args):
         """Fetches Quote by ID, usage: !quote get <id>"""
+        if len(args) != 1:
+            return "Usage: !quote get <id>"
+
         self.cur.execute('''select * from quotes where id = ?''', (args,))
         quote = self.cur.fetchone()
         if quote is None:
-            ret = 'Did not found a quote with id = %d.' % int(args)
+            return 'Did not found a quote with id = %d.' % int(args)
         else:
-            ret = '[%d] %s' % (quote[0], quote[1])
-        self.con.commit()
-        return ret
+            return '[%d] %s' % (quote[0], quote[1])
